Abstract out common elements in *CommentReportEntity
